requirements.txt
absl-py==2.1.0
annotated-types==0.5.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.0.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
cycler==0.11.0
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
deskew==1.0.26
dnspython==2.3.0
docopt==0.6.2
easyocr==1.7.1
editdistance==0.6.2
efficientnet==1.0.0
email-validator==2.0.0.post2
entrypoints==0.4
essential-generators==1.0
exceptiongroup==1.2.0
fast-deskew==1.0
fastapi==0.103.2
fastjsonschema==2.19.1
flatbuffers==23.5.26
fonttools==4.38.0
gast==0.4.0
google-auth==2.28.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.0
h11==0.14.0
h5py==3.8.0
hdf5storage==0.1.19
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.6
imageio==2.31.2
imgaug==0.4.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab_widgets==3.0.10
keras==2.11.0
Keras-Applications==1.0.8
keras-ocr==0.8.9
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==3.0.2
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.6.0
networkx==2.6.3
ninja==1.11.1.1
notebook==6.5.6
notebook_shim==0.2.4
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
orjson==3.9.7
packaging==23.2
pandas==1.3.5
pandocfilters==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.17.1
prompt-toolkit==3.0.43
protobuf==3.19.6
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycparser==2.21
pydantic==2.5.3
pydantic-extra-types==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.14.6
Pygments==2.17.2
pyparsing==3.1.1
pyrsistent==0.19.3
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.8
pytz==2024.1
PyWavelets==1.3.0
pywin32==306
pywinpty==2.0.10
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.4
QtPy==2.4.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.5.0
Send2Trash==1.8.2
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.4.1
starlette==0.27.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
tifffile==2021.11.2
tinycss2==1.2.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.66.2
traitlets==5.9.0
typing_extensions==4.7.1
ujson==5.7.0
urllib3==2.0.7
uvicorn==0.22.0
validators==0.20.0
watchfiles==0.20.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.2.3
widgetsnbextension==4.0.10
wrapt==1.16.0
zipp==3.15.0
